requirements.txt
Flask==3.1.0
Flask_Limiter==3.9.2
langchain==0.3.13
langchain_openai==0.2.13
openai==1.58.1
pdfplumber==0.11.4
Werkzeug==3.1.3
python-dotenv
